# tests/snapshots/openapi_routers.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "guguwebui routers snapshot",
    "version": "test"
  },
  "paths": {
...
    "/api/chat/verifications/{code}": {
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "Chat Check Verification",
        "description": "检查验证码验证状态：data.verified 为业务状态（不再复用 status）",
        "operationId": "chat_check_verification___api_chat_verifications_{code}__GET",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chat/accounts/{name}/password": {
      "put": {
        "tags": [
          "chat"
        ],
        "summary": "Chat Set Password",
        "description": "设置聊天页用户密码（name 必须与验证码绑定的玩家一致；成功后直接签发会话）",
        "operationId": "chat_set_password___api_chat_accounts_{name}_password__PUT",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatSetPasswordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chat/sessions": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Chat Login",
        "description": "聊天页用户登录（创建会话；失败次数过多 → 429 too_many_attempts）",
        "operationId": "chat_login___api_chat_sessions__POST",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chat/session/{session_id}": {
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "Chat Check Session",
        "description": "检查聊天页会话状态（不存在 → 404 session_not_found，过期 → 401 session_expired）",
        "operationId": "chat_check_session___api_chat_session_{session_id}__GET",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "chat"
        ],
        "summary": "Chat Logout",
        "description": "聊天页用户退出登录",
        "operationId": "chat_logout___api_chat_session_{session_id}__DELETE",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chat/messages": {
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "Get Chat Messages",
        "description": "获取聊天消息（新→旧；after_id/before_id 为游标，与 offset 二选一）\n\n权限：WebUI 登录态；或公开聊天页启用时的有效聊天会话（query session_id）。",
        "operationId": "get_chat_messages___api_chat_messages__GET",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "after_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "After Id"
            }
          },
          {
            "name": "before_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Before Id"
            }
          },
          {
            "name": "session_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "公开聊天页会话 ID（未登录 WebUI 时必填）",
              "title": "Session Id"
            },
            "description": "公开聊天页会话 ID（未登录 WebUI 时必填）"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "chat"
        ],
        "summary": "Chat Clear Messages",
        "description": "清空聊天消息（仅管理员）",
        "operationId": "chat_clear_messages___api_chat_messages__DELETE",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessEnvelope"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Send Chat Message",
        "description": "发送聊天消息到游戏（公开接口；WebUI 登录用户且用户名与 player_id 一致时走管理员通道）",
        "operationId": "send_chat_message___api_chat_messages__POST",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatMessageCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chat/messages/incremental": {
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "Get New Chat Messages",
        "description": "获取新消息与在线状态（轮询接口，基于最后消息 ID；player_id 作心跳）\n\n聊天会话访客的心跳固定记为会话绑定的玩家，不能替他人上报在线。",
        "operationId": "get_new_chat_messages___api_chat_messages_incremental__GET",
        "parameters": [
          {
            "name": "after_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "After Id"
            }
          },
          {
            "name": "player_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Player Id"
            }
          },
          {
            "name": "session_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "公开聊天页会话 ID（未登录 WebUI 时必填）",
              "title": "Session Id"
            },
            "description": "公开聊天页会话 ID（未登录 WebUI 时必填）"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiSuccessEnvelope": {
        "properties": {
          "status": {
            "type": "string",
            "const": "success",
            "title": "Status",
            "default": "success"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "data": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Data"
          }
        },
        "type": "object",
        "title": "ApiSuccessEnvelope",
        "description": "统一成功外壳（2.3）：status/message?/data?。"
      },
...
      "ChatLoginRequest": {
        "properties": {
          "player_id": {
            "type": "string",
            "title": "Player Id"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "player_id",
          "password"
        ],
        "title": "ChatLoginRequest",
        "description": "聊天页用户登录"
      },
      "ChatMessageCreateRequest": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "player_id": {
            "type": "string",
            "title": "Player Id"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          }
        },
        "type": "object",
        "required": [
          "message",
          "player_id"
        ],
        "title": "ChatMessageCreateRequest",
        "description": "发送聊天消息（管理员通道可省略 session_id）"
      },
      "ChatSetPasswordRequest": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "code",
          "password"
        ],
        "title": "ChatSetPasswordRequest",
        "description": "设置聊天页用户密码（code 为游戏内验证码，路径 name 须与其绑定玩家一致）"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PageData": {
        "properties": {
          "items": {
            "items": {},
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "offset"
        ],
        "title": "PageData",
        "description": "统一分页负载（2.3）：data = {items, total, offset, limit}。"
      },
      "PageEnvelope": {
        "properties": {
          "status": {
            "type": "string",
            "const": "success",
            "title": "Status",
            "default": "success"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PageData"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "PageEnvelope",
        "description": "分页成功外壳。"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}